% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{alpha_graph}
\alias{alpha_graph}
\title{Tau vs Alpha sensitivity analysis plot}
\usage{
alpha_graph(alpha_grid = NULL, tau_star, tau_star_se, ci_sig = 0.05)
}
\arguments{
\item{alpha_grid}{a grid to loop over alpha (if not provided will create)}

\item{tau_star}{the estimated treatment effect}

\item{tau_star_se}{the estimated treatment effect standard errors}

\item{ci_sig}{confidence interval significance level for rejection of H0: tau= 0 (default is 5%)}
}
\description{
Tau vs Alpha sensitivity analysis plot
}
