% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_sens.R
\name{alpha_loop}
\alias{alpha_loop}
\title{Alpha grid search}
\usage{
alpha_loop(alpha_grid = NULL, tau_star, tau_star_se, ci_sig = 0.05)
}
\arguments{
\item{alpha_grid}{a grid to loop over alpha (if not provided will create with 101 values)}

\item{tau_star}{the estimated treatment effect}

\item{tau_star_se}{the estimated treatment effect standard errors}

\item{ci_sig}{confidence interval signifcance level for rejection of H0: tau= 0 (default is 5%)}
}
\value{
a dataframe containing alpha levels and corresponding taus and rejections of null hypothesis
}
\description{
Alpha grid search
}
