% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_sens.R
\name{disc_test}
\alias{disc_test}
\title{Conservative Continuity Hypothesis Testing}
\usage{
disc_test(alpha_grid = NULL, tau_star, tau_star_se, ci_sig = 0.05)
}
\arguments{
\item{alpha_grid}{a grid to loop over alpha (if not provided will create with 101 values)}

\item{tau_star}{the estimated treatment effect}

\item{tau_star_se}{the estimated treatment effect standard errors}

\item{ci_sig}{confidence interval signifcance level for rejection of H0: tau= 0 (default is 5%)}
}
\value{
A list containing alpha-max, tau(alpha-max) p-value of rejection of the conservative H0 was rejected
}
\description{
This function tests the hypothesis of treatment effect equalling maximum ``continuous'' treatment effect
}
