% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{rd_graph}
\alias{rd_graph}
\title{RD graph (outcome vs. running)}
\usage{
rd_graph(dt, c)
}
\arguments{
\item{c}{RD cutoff}

\item{A}{dataframe, where "Y" is the outcome, "D" is running var., W is treatment var.}
}
\description{
RD graph (outcome vs. running)
}
