% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_tab}
\alias{summary_tab}
\title{Cosmetic function for autmatic latex table saving using kableExtra}
\usage{
summary_tab(data)
}
\arguments{
\item{data}{Table dataframe}
}
\value{
Latex table output
}
\description{
Cosmetic function for autmatic latex table saving using kableExtra
}
