% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{threeway_graph}
\alias{threeway_graph}
\title{A cosemtic function to organize three graphs in one row}
\usage{
threeway_graph(g1, g2, g3)
}
\arguments{
\item{g1}{Left graph}

\item{g2}{Center graph}

\item{g3}{Right graph}
}
\value{
A single graph
}
\description{
A cosemtic function to organize three graphs in one row
}
