% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{threeway_graph_non_title}
\alias{threeway_graph_non_title}
\title{A cosemtic function to organize three graphs in one row, without a title}
\usage{
threeway_graph_non_title(g1, g2, g3)
}
\arguments{
\item{g1}{Left graph}

\item{g2}{Center graph}

\item{g3}{Right graph}
}
\value{
A single graph
}
\description{
A cosemtic function to organize three graphs in one row, without a title
}
